requirements.txt
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.0
prefect~=2.17.1
prefect-sqlalchemy==0.4.0
pyarrow==15.0.0
pylint==3.0.3
pymongo==4.6.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
fastapi~=0.110.1
setuptools~=68.2.0